+++ src/index.tsx
@@ -1,5 +1,6 @@
 export * from './Fetch';
 export * from './FetchConfigContext';
+export * from './stores/CacheStore';
 export * from './stores/IndexedDBStore';
 export * from './stores/MemoryStore';
 export * from './stores/StorageStore';